Fixes the mapping of soil properties and surface area
The current pf-to-elm map is only applicable for the exchnage of
liquid mass between the two models, but the map is not valid for
mapping soil properties and surface area. Instead of introducing
new maps, the weight matrix of the existing map is used to correctly
map the soil properties and surface area.